add comments, update logic

+++ src/main/resources/db/migration/V1__create_processors_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `processors` (
-    `id` int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint(11) NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name` varchar(100) NOT NULL UNIQUE,
     `manufacturer_name` varchar(100) DEFAULT NULL,
     `socket` varchar(100) DEFAULT NULL,
